Move `PpdbBigQueryConfig` into `bigquery.ppdb_bigquery` module
It doesn't really need its own standalone module.
